More File IO Progress
Added a test class for file input/output.

Removed fromSaveString function from IGraphElement, instead loading is handled by GraphReader.

The data format for Graphs is almost certainly going to change, thus the included test files will likely break in future versions.

Added clearElements and addElements functions to Graph2D for handling large data changes.

Also made various changes to ProgramView to support file IO.

+++ src/model/FieldArrow.java
@@ -13,7 +13,7 @@ import utils.Vec2D;
 // beyond being a distinction for the view model
 public class FieldArrow extends Arrow2D
 {
-	public FieldArrow(int xPos, int yPos, int xMag, int yMag)
+	public FieldArrow(double xPos, double yPos, double xMag, double yMag)
 	{
 		super(xPos, yPos, xMag, yMag);
 	}
